Solved part 1!
To solve followoing parts it's a good idea to make functions as general as possible I notice. It started to ask for five leading zeros, I now have to alter the function to a 6 zeros funciton. This requiers some easy coding chaning a one variable. But I could have coded the function to be more flexible from the get go... Will consider for the future.